+++ src/cli.ts
@@ -1,35 +1,31 @@
 #!/usr/bin/env node
 
-import commander from 'commander';
-import * as fs from 'fs';
-import * as path from 'path';
+import {program as commander} from 'commander';
 
-import {deleteAllClients} from './commands/deleteAllClients';
-import {resetPassword} from './commands/resetPassword';
-import {setName} from './commands/setName';
-import {getAllClients, updateClient} from './commands/updateOrGetClient';
-import {CommonOptions} from './CommonOptions';
-import {getLogger, init, tryAndExit} from './util';
+import {
+  deleteAllClients,
+  deleteMessage,
+  getAllClients,
+  getSelf,
+  resetPassword,
+  sendMessage,
+  setAvailabilityStatus,
+  setName,
+  updateClient,
+} from './commands/';
+import {setEmail} from './commands/setEmail';
+import {addHTTPS, getLogger, init, tryAndExit, getPackageJson} from './util';
 
-const defaultPackageJsonPath = path.join(__dirname, 'package.json');
-const packageJsonPath = fs.existsSync(defaultPackageJsonPath)
-  ? defaultPackageJsonPath
-  : path.join(__dirname, '../package.json');
-
-const packageJson = fs.readFileSync(packageJsonPath, 'utf-8');
-const {description, name, version} = JSON.parse(packageJson);
+const {bin, description, version} = getPackageJson();
 const commanderOptions = commander.opts();
-
-const defaultOptions: CommonOptions = {
-  defaultBackendURL: 'https://staging-nginz-https.zinfra.io',
-};
+const defaultBackendURL = 'staging-nginz-https.zinfra.io';
 
 const logger = getLogger('cli');
 
 commander
-  .name(name.replace(/^@.+\//, ''))
+  .name(Object.keys(bin)[0])
   .description(description)
-  .option('-b, --backend <URL>', `specify the Wire backend URL (default: "${defaultOptions.defaultBackendURL}")`)
+  .option('-b, --backend <URL>', 'specify the Wire backend URL')
   .option('-d, --dry-run', `don't send any data (beside logging in and out)`)
   .option('-e, --email <address>', 'specify your Wire email address')
   .option('-p, --password <password>', 'specify your Wire password')
@@ -42,7 +38,7 @@ commander
 commander
   .command('delete-all-clients')
   .description('delete all clients')
-  .option('-b, --backend <URL>', `specify the Wire backend URL (default: "${defaultOptions.defaultBackendURL}")`)
+  .option('-b, --backend <URL>', 'specify the Wire backend URL')
   .option('-d, --dry-run', `don't send any data (beside logging in and out)`)
   .option('-e, --email <address>', 'specify your email address')
   .option('-p, --password <password>', 'specify your Wire password')
@@ -50,11 +46,11 @@ commander
     tryAndExit(() => {
       init();
       return deleteAllClients({
-        ...defaultOptions,
-        ...(commanderOptions?.backend && {backendURL: commanderOptions.backend}),
-        ...(commanderOptions?.dryRun && {dryRun: commanderOptions.dryRun}),
-        ...(commanderOptions?.email && {emailAddress: commanderOptions.email}),
-        ...(commanderOptions?.password && {password: commanderOptions.password}),
+        backendURL: addHTTPS(commanderOptions.backend),
+        defaultBackendURL,
+        dryRun: commanderOptions.dryRun,
+        emailAddress: commanderOptions.email,
+        password: commanderOptions.password,
       });
     })
   );
@@ -62,23 +58,23 @@ commander
 commander
   .command('set-client-label')
   .description(`update a client's label`)
-  .option('-b, --backend <URL>', `specify the Wire backend URL (default: "${defaultOptions.defaultBackendURL}")`)
+  .option('-b, --backend <URL>', 'specify the Wire backend URL')
   .option('-d, --dry-run', `don't send any data (beside logging in and out)`)
   .option('-e, --email <address>', 'specify your email address')
   .option('-i, --client-id <id>', `specify the client's ID`)
   .option('-l, --label <label>', 'specify the new label')
-  .option('-t, --password <password>', 'specify your Wire password')
-  .action(() =>
+  .option('-p, --password <password>', 'specify your Wire password')
+  .action((localOptions: {clientId?: string; label?: string} | undefined) =>
     tryAndExit(() => {
       init();
       return updateClient({
-        ...defaultOptions,
-        ...(commanderOptions?.clientId && {clientId: commanderOptions.clientId}),
-        ...(commanderOptions?.label && {label: commanderOptions.label}),
-        ...(commanderOptions?.backend && {backendURL: commanderOptions.backend}),
-        ...(commanderOptions?.dryRun && {dryRun: commanderOptions.dryRun}),
-        ...(commanderOptions?.email && {emailAddress: commanderOptions.email}),
-        ...(commanderOptions?.password && {password: commanderOptions.password}),
+        backendURL: addHTTPS(commanderOptions.backend),
+        clientId: localOptions?.clientId!,
+        defaultBackendURL,
+        dryRun: commanderOptions.dryRun,
+        emailAddress: commanderOptions.email,
+        label: localOptions?.label!,
+        password: commanderOptions.password,
       });
     })
   );
@@ -86,7 +82,7 @@ commander
 commander
   .command('reset-password')
   .description('reset your password')
-  .option('-b, --backend <URL>', `specify the Wire backend URL (default: "${defaultOptions.defaultBackendURL}")`)
+  .option('-b, --backend <URL>', 'specify the Wire backend URL')
   .option('-c, --continue', 'skip initiation (if you already received the password reset email)')
   .option('-d, --dry-run', `don't send any data (beside logging in and out)`)
   .option('-e, --email <address>', 'specify your email address')
@@ -94,11 +90,11 @@ commander
     tryAndExit(() => {
       init();
       return resetPassword({
-        ...defaultOptions,
-        ...(commanderOptions?.backend && {backendURL: commanderOptions.backend}),
-        ...(commanderOptions?.continue && {skipInitation: commanderOptions.continue}),
-        ...(commanderOptions?.dryRun && {dryRun: commanderOptions.dryRun}),
-        ...(commanderOptions?.email && {emailAddress: commanderOptions.email}),
+        backendURL: addHTTPS(commanderOptions.backend),
+        defaultBackendURL,
+        dryRun: commanderOptions.dryRun,
+        emailAddress: commanderOptions.email,
+        skipInitation: commanderOptions.continue,
       });
     })
   );
@@ -106,7 +102,7 @@ commander
 commander
   .command('get-all-clients')
   .description('get all clients data')
-  .option('-b, --backend <URL>', `specify the Wire backend URL (default: "${defaultOptions.defaultBackendURL}")`)
+  .option('-b, --backend <URL>', 'specify the Wire backend URL')
   .option('-d, --dry-run', `don't send any data (beside logging in and out)`)
   .option('-e, --email <address>', 'specify your email address')
   .option('-p, --password <password>', 'specify your Wire password')
@@ -114,11 +110,31 @@ commander
     tryAndExit(() => {
       init();
       return getAllClients({
-        ...defaultOptions,
-        ...(commanderOptions?.backend && {backendURL: commanderOptions.backend}),
-        ...(commanderOptions?.dryRun && {dryRun: commanderOptions.dryRun}),
-        ...(commanderOptions?.email && {emailAddress: commanderOptions.email}),
-        ...(commanderOptions?.password && {password: commanderOptions.password}),
+        backendURL: addHTTPS(commanderOptions.backend),
+        defaultBackendURL,
+        dryRun: commanderOptions.dryRun,
+        emailAddress: commanderOptions.email,
+        password: commanderOptions.password,
+      });
+    })
+  );
+
+commander
+  .command('get-self')
+  .description('get self data')
+  .option('-b, --backend <URL>', 'specify the Wire backend URL')
+  .option('-d, --dry-run', `don't send any data (beside logging in and out)`)
+  .option('-e, --email <address>', 'specify your email address')
+  .option('-p, --password <password>', 'specify your Wire password')
+  .action(() =>
+    tryAndExit(() => {
+      init();
+      return getSelf({
+        backendURL: addHTTPS(commanderOptions.backend),
+        defaultBackendURL,
+        dryRun: commanderOptions.dryRun,
+        emailAddress: commanderOptions.email,
+        password: commanderOptions.password,
       });
     })
   );
@@ -126,21 +142,113 @@ commander
 commander
   .command('set-name')
   .description('set your name')
-  .option('-b, --backend <URL>', `specify the Wire backend URL (default: "${defaultOptions.defaultBackendURL}")`)
+  .option('-b, --backend <URL>', 'specify the Wire backend URL')
   .option('-d, --dry-run', `don't send any data (beside logging in and out)`)
   .option('-e, --email <address>', 'specify your email address')
   .option('-n, --new-name <name>', 'specify your new name')
   .option('-p, --password <password>', 'specify your Wire password')
-  .action(() =>
+  .action((localOptions: {newName?: string} | undefined) =>
     tryAndExit(() => {
       init();
       return setName({
-        ...defaultOptions,
-        ...(commanderOptions?.backend && {backendURL: commanderOptions.backend}),
-        ...(commanderOptions?.dryRun && {dryRun: commanderOptions.dryRun}),
-        ...(commanderOptions?.email && {emailAddress: commanderOptions.email}),
-        ...(commanderOptions?.newName && {name: commanderOptions.newName}),
-        ...(commanderOptions?.password && {password: commanderOptions.password}),
+        backendURL: addHTTPS(commanderOptions.backend),
+        defaultBackendURL,
+        dryRun: commanderOptions.dryRun,
+        emailAddress: commanderOptions.email,
+        newName: localOptions?.newName,
+        password: commanderOptions.password,
+      });
+    })
+  );
+
+commander
+  .command('set-email')
+  .description('set your email address')
+  .option('-b, --backend <URL>', 'specify the Wire backend URL')
+  .option('-d, --dry-run', `don't send any data (beside logging in and out)`)
+  .option('-e, --email <address>', 'specify your email address')
+  .option('-n, --new-email <address>', 'specify your new email address')
+  .option('-p, --password <password>', 'specify your Wire password')
+  .action((localOptions: {newEmail?: string} | undefined) =>
+    tryAndExit(() => {
+      init();
+      return setEmail({
+        backendURL: addHTTPS(commanderOptions.backend),
+        defaultBackendURL,
+        dryRun: commanderOptions.dryRun,
+        emailAddress: commanderOptions.email,
+        newEmailAddress: localOptions?.newEmail,
+        password: commanderOptions.password,
+      });
+    })
+  );
+
+commander
+  .command('set-availability')
+  .description('set your availability status')
+  .option('-b, --backend <URL>', 'specify the Wire backend URL')
+  .option('-d, --dry-run', `don't send any data (beside logging in and out)`)
+  .option('-e, --email <address>', 'specify your Wire email address')
+  .option('-p, --password <password>', 'specify your Wire password')
+  .option('-s, --status <status>', 'specify the status type to be set (0/1/2/3 or none/available/away/busy)')
+  .action((localOptions: {status?: string} | undefined) =>
+    tryAndExit(() => {
+      init();
+      return setAvailabilityStatus({
+        backendURL: addHTTPS(commanderOptions.backend),
+        defaultBackendURL,
+        dryRun: commanderOptions.dryRun,
+        emailAddress: commanderOptions.email,
+        password: commanderOptions.password,
+        statusType: localOptions?.status,
+      });
+    })
+  );
+
+commander
+  .command('send')
+  .description('send a message')
+  .option('-b, --backend <URL>', 'specify the Wire backend URL')
+  .option('-d, --dry-run', `don't send any data (beside logging in and out)`)
+  .option('-e, --email <address>', 'specify your Wire email address')
+  .option('-p, --password <password>', 'specify your Wire password')
+  .option('-c, --conversation <ID>', 'specify the conversation for the message to be sent to')
+  .option('-m, --message <text>', 'specify the message to be sent')
+  .action((localOptions: {conversation?: string; message?: string} | undefined) =>
+    tryAndExit(() => {
+      init();
+      return sendMessage({
+        backendURL: addHTTPS(commanderOptions.backend),
+        conversationID: localOptions?.conversation,
+        defaultBackendURL,
+        dryRun: commanderOptions.dryRun,
+        emailAddress: commanderOptions.email,
+        message: localOptions?.message,
+        password: commanderOptions.password,
+      });
+    })
+  );
+
+commander
+  .command('delete')
+  .description('delete a message')
+  .option('-b, --backend <URL>', 'specify the Wire backend URL')
+  .option('-c, --conversation <ID>', 'specify the conversation in which the message should be deleted')
+  .option('-d, --dry-run', `don't send any data (beside logging in and out)`)
+  .option('-e, --email <address>', 'specify your Wire email address')
+  .option('-m, --message <ID>', 'specify the message ID to be deleted')
+  .option('-p, --password <password>', 'specify your Wire password')
+  .action((localOptions: {conversation?: string; message?: string} | undefined) =>
+    tryAndExit(() => {
+      init();
+      return deleteMessage({
+        backendURL: addHTTPS(commanderOptions.backend),
+        conversationID: localOptions?.conversation,
+        defaultBackendURL,
+        dryRun: commanderOptions.dryRun,
+        emailAddress: commanderOptions.email,
+        messageID: localOptions?.message,
+        password: commanderOptions.password,
       });
     })
   );