+++ src/commands/getSelf.ts
@@ -1,38 +1,21 @@
-import prompts from 'prompts';
 import {APIClient} from '../APIClient';
 import {CommonOptions} from '../CommonOptions';
 import {getBackendURL, getEmailAddress, getLogger, getPassword} from '../util';
 
-export interface SetNameOptions extends CommonOptions {
-  newName?: string;
-}
+export interface GetSelfOptions extends CommonOptions {}
 
-const logger = getLogger('set-name');
+const logger = getLogger('get-self');
 
-export async function setName({
+export async function getSelf({
   backendURL,
   defaultBackendURL,
   dryRun,
   emailAddress,
-  newName,
   password,
-}: SetNameOptions): Promise<void> {
+}: GetSelfOptions): Promise<void> {
   backendURL ||= await getBackendURL(defaultBackendURL);
   emailAddress ||= await getEmailAddress();
   password ||= await getPassword();
-  newName ||= (
-    await prompts(
-      {
-        max: 128,
-        message: 'Enter your new name (max. 128 chars):',
-        name: 'newName',
-        type: 'text',
-      },
-      {
-        onCancel: () => process.exit(),
-      }
-    )
-  ).newName;
 
   const apiClient = new APIClient(backendURL, emailAddress, password);
 
@@ -40,10 +23,9 @@ export async function setName({
 
   await apiClient.login();
 
-  logger.info(`Setting new name to "${newName}" ...`);
-
   if (!dryRun) {
-    await apiClient.putSelf({name: newName!});
+    const response = await apiClient.getSelf();
+    logger.info(response.data);
   }
 
   logger.info('Logging out ...');